jsファイル作成しturbolinkより上部に配置
